Video #25 completed

+++ app/Http/Controllers/TwitController.php
@@ -17,6 +17,10 @@ class TwitController extends Controller
             'content' => 'required|min:10|max:280'
         ]);
 
+        // Save current logged-in user
+        $validated['user_id'] = auth()->user()->id;
+
+        // Save twit
         Twittah::create($validated);
 
         return redirect()->route('dashboard')->with('success', 'Twit create successfully');
@@ -24,12 +28,20 @@ class TwitController extends Controller
 
     public function edit(Twittah $twit) {
 
+        if(auth()->id() !== $twit->user_id) {
+            abort(404);
+        }
+
         $editing = true;
 
         return view('twits.show', compact('twit', 'editing'));
     }
 
     public function update(Twittah $twit) {
+
+        if(auth()->id() !== $twit->user_id) {
+            abort(404);
+        }
 
         $validated = request()->validate([
             'content' => 'required|min:10|max:280'
@@ -42,6 +54,10 @@ class TwitController extends Controller
 
     // Route model binding
     public function destroy(Twittah $id) {
+
+        if(auth()->id() !== $id->user_id) {
+            abort(404);
+        }
 
         $id->delete();
